% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_backend.R
\name{backend.base}
\alias{backend.base}
\title{An incomplete backend.}
\usage{
backend.base(X, Y, cache.size = 0, n.eff = nrow(X))
}
\arguments{
\item{X}{Matrix. The data vector for the X side}

\item{Y}{Matrix. The data vector for the Y side}

\item{cache.size}{The cache size storing the correlations. Defaults to 0 (don't store anything).}

\item{n.eff}{Effective sample size. Nothing to change if 
rows of X and Y are independent.}
}
\description{
Other backends build on this. See and use backend instead.
}
\keyword{internal}
