% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R, R/perm_backend.R
\name{backend}
\alias{backend}
\alias{backend.perm}
\title{Backends : engines for p-value computation.}
\usage{
backend.perm(X, Y, cache.size = 0)
}
\arguments{
\item{X}{Matrix. The data vector for the X side}

\item{Y}{Matrix. The data vector for the Y side}

\item{cache.size}{The cache size storing the correlations. Defaults to 0 (don't store anything).}
}
\description{
A backend is an S3 class which implments the methods \code{\link{pvals}}, \code{\link{cors}} and \code{\link{pvals_singleton}}. Different backends provide p-value computation under different scenarios (e.g. squared vs sum of squared correlations, Indpendence within Gene and SNP's) and may do complicated precomputations on the data matrices (\code{X} and \code{Y}) behind the scenes. Yet all of them provide a unified interface.
The methods below are constructors for the differnt backend objects and return an S3 object which implements pvalue methods listed above.
}
\section{Functions}{
\itemize{
\item \code{backend.perm}: Constructor for sum of squared correlations under independent Gene and SNP sets (the strong null).
}}

\examples{
\dontrun{
 X <- matrix(rnorm(100), nrow=10)
 Y <- matrix(rnorm(50), nrow=10)
 bk <- backend.indepNormal(X, Y)
 #Calculate Y pvalues againt the X (column) subset {1,2..,7}
 py <- pvals(bk, 1:7)
 #Calculate X pvalues againt the Y (column) subset {2, 3}. Use global numbering.
 px <- pvals(bk, 10 + c(2,3))
}
}
