% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{cors}
\alias{cors}
\title{Calculate correlation vector from the set A}
\usage{
cors(bk, A)
}
\arguments{
\item{bk}{An object of class \link{backend}}

\item{A}{Either a subset of X or Y columns (in global numbering)}
}
\description{
The backends typically have a faster way to implement this (using precomputations)
but this is roughly equivalent to \code{cor(X, Y[,A-dx])} if \code{min(A) > dx} else it is equal to \code{cor(Y, X[,A])}.
}
