% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_none}
\alias{diagnostics_none}
\title{Don't collect any diagnostics from the method.}
\usage{
diagnostics_none(event, e = parent.frame())
}
\description{
Don't collect any diagnostics from the method.
}
\section{Functions}{
\itemize{
\item \code{diagnostics_none}: 
}}

\keyword{internal}
