% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{This is the default diagnositcs method passed to cbce.}
\usage{
diagnostics(event, e = parent.frame())
}
\description{
It collects data about the internal state of the method 
and returns it alonside the result.
}
\details{
Currently, the diagnostics collected are:
\enumerate{
\item Consecutive_jaccards : The jaccard distance at each stage
 \item sizes : The sizes at each stage in the extaction
 \item disjointed : Whether disjointed at each point.
\item cycle : Whether cycled at each point.
\item Extraction time.
}
}
\keyword{internal}
