% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_and_summarize}
\alias{filter_and_summarize}
\title{Filter and summarize the results}
\usage{
filter_and_summarize(
  extract_res,
  plot.dendrogram = FALSE,
  hclust.method = "average",
  logpval.thresh = 0
)
}
\arguments{
\item{extract_res}{\code{result$extract_res} where 
\code{result} is the returned by the \code{\link{cbce}} 
procedure}

\item{plot.dendrogram}{logical Plot the dendrogram 
along with the line it is cut at.}

\item{hclust.method}{The clustering method to use 
(passed to hclust)}
}
\value{
A data frame, each row of which represents the summary of a filtered bimodule. 
The columns of the frame are:
\itemize{
 \item{index} This is the index of the bimodule in extract_res.
 \item{x.size, y.size} These are the sizes of the X, Y set of the bimodules
 \item{score} The score assigned to the bimodule based on how strong the
 correlation is (vs. expected). 
 \item{group.size} This bimodule is a representative for a group of bimodules of this size.
}
}
\description{
The communities are clustered based on their jaccard 
distances and the dendrogram is cut to obtain the same
number of communities as are obtained by the 
effective-count formula. For each of the cut 
subtree an appropriate bimodule is chosen to represent it.
}
\examples{
\dontrun{
n <- 100
dx <- 50
dy <- 70

X <- matrix(rnorm(n*dx), ncol=dx)
Y <- matrix(rnorm(n*dy), ncol=dy)
res <- cbce2(X, Y)

df <- filter_and_summarize(res$extract_res)
# or just df <- res$filtered_result.df

# The filtered bimodules:
bms <- rlist::list.map(res$extract_res[df$index], bimod) 
}

}
\keyword{internal}
