% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_bimodules}
\alias{filter_bimodules}
\title{Filter bimodules for results that look similar.}
\usage{
filter_bimodules(bms, hclust.method = "average", plot.dendrogram = FALSE)
}
\arguments{
\item{bms}{a list of bimodules. Each bimodule should be 
a list with two elements x and y.}

\item{hclust.method}{The clustering method to use 
(passed to hclust)}

\item{plot.dendrogram}{logical Plot the dendrogram 
along with the line it is cut at.}
}
\value{
The list of filtered bimodules. If there are 
ties the first one in order is preferred.
}
\description{
The bimodules are clustered based on their jaccard 
distances and the dendrogram is cut to obtain the same
number of bimodules as are obtained by the 
effective-count formula. For each of the cut 
subtree an appropriate bimodule is chosen to represent it.
}
