% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_parameter.R
\name{half_permutation_fdr}
\alias{half_permutation_fdr}
\title{Estimate the false discovery rate for the method across a 
variety of parameters.}
\usage{
half_permutation_fdr(
  X,
  Y,
  alphas,
  num.sims,
  method = cbce.fast,
  cov = NULL,
  fdr = "all.pairs"
)
}
\arguments{
\item{X, Y, cov}{The original data that we intend to run the method 
on.}

\item{alphas}{List of parameters to run the method on.}

\item{num.sims}{The number of simulation to evaluate over.}

\item{method}{The method to run. The method should take 
arguments \code{X, Y, cov, alpha} and return 
a collection of bi-modules.}

\item{fdr}{The type of false discovery metric to use. Can be one of 'all.pairs' or 'imp.pairs'.}

\item{timeout}{}
}
\value{
Returns a numeric vector of size length(alphas) that have the fdrs.
}
\description{
For each parameter value, generate \code{num.sims} half-permuted 
datasets, and estimate the half-permutation FDR.
}
\keyword{internal}
