% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{init}
\alias{init}
\title{Initialization function}
\usage{
init(p, indx, alpha, init_method)
}
\arguments{
\item{p}{backend}

\item{indx}{The node (in global numbering to initialize from)}

\item{alpha}{The cutoff to use for pvalues.}

\item{init_method}{One of \code{c("conservative-BH", 
"non-conservative-BH", "BH-0.5", "no-multiple-testing")}
        gives the type of initializtion scheme.}
}
\description{
Initialization function
}
\keyword{internal}
