% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_cli}
\alias{interaction_cli}
\title{Provides a command line progress indicator for \code{\link{cbce}}
Just pass \code{interaction=interaction_cli} as an argument to cbce.}
\usage{
interaction_cli(event, env = parent.frame())
}
\arguments{
\item{event}{The name of the event of the callback}

\item{env}{The environment to access the method internals and store
persistent state.}
}
\description{
Provides a command line progress indicator for \code{\link{cbce}}
Just pass \code{interaction=interaction_cli} as an argument to cbce.
}
