% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R, R/perm_backend.R
\name{pvals}
\alias{pvals}
\alias{pvals.perm}
\title{Calculate p-values against a testing set.}
\usage{
pvals(bk, B, thresh.alpha = 1)

\method{pvals}{perm}(bk, B, thresh.alpha = 1)
}
\arguments{
\item{bk}{An object of class backend}

\item{B}{A set of either X or Y indices (using global numbering). The method calculates the pvalues from this set to the opposite side.}

\item{thresh.alpha}{We are only interested in p-values less than thresh.alpha. So the method free to return NA for pvalues > thresh.alpha if it helps with optimization.}
}
\value{
The result is a vector of p-values of length ncol(X) or ncol(Y) depending on whether B is from Y or X respectively.
}
\description{
Given a testing set B, return the p-values for the opposite side.
This is an S3 method that is implemented by all backends.
}
\section{Methods (by class)}{
\itemize{
\item \code{perm}: Implementation for sum of squared correlations under independent Gene and SNP sets. 
Uses the permutation moments approximation.
}}

\seealso{
\code{\link{pvals_singleton}}
}
