% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-pval.R
\name{summary_pval}
\alias{summary_pval}
\title{Compute the score (i.e a summarizing p-vlaue) for the 
bimodule given by the entire X and Y matrices.}
\usage{
summary_pval(X, Y, log.p = TRUE, summary = NULL, n.eff = nrow(X))
}
\description{
Compute the score (i.e a summarizing p-vlaue) for the 
bimodule given by the entire X and Y matrices.
}
\keyword{internal}
