% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_cli}
\alias{interaction_cli}
\title{Provides a CLI for interaction with cbce.}
\usage{
interaction_cli(event, env = parent.frame())
}
\description{
Pass interaction=\code{interaction_cli} to get a progress indicator.
}
\keyword{internal}
