% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_gui}
\alias{interaction_gui}
\title{Provides a GUI for interaction with cbce.}
\usage{
interaction_gui(...)
}
\description{
Pass interaction=\code{interaction_gui} to cbce for a GUI interface.
The global variable \code{TITLE} constrols the title of the GUI.
}
\keyword{internal}
