% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R, R/perm_backend.R
\name{pvals}
\alias{pvals}
\alias{pvals.perm}
\title{Backends : engines for p-value computation.}
\usage{
pvals(bk, B, thresh.alpha = 1)

\method{pvals}{perm}(bk, B, thresh.alpha = 1)
}
\arguments{
\item{bk}{An object of class backend}

\item{B}{A set of either X or Y indices (using global numbering). The method calculates the pvalues from this set to the opposite side.}

\item{thresh.alpha}{We are only interested in p-values less than thresh.alpha. So the method free to return NA for pvalues > thresh.alpha if it helps with optimization.}
}
\value{
The result is a vector of p-values of length ncol(X) or ncol(Y) depending on whether B is from Y or X respectively.
}
\description{
A backend is an S3 class which implments the methods 
\code{\link{pvals}}, \code{\link{cors}} and 
\code{\link{pvals_singleton}}. Different backends provide 
p-value computation under different scenarios 
(e.g. squared vs sum of squared correlations, 
Indpendence within Gene and SNP's) and may do complicated
 precomputations on the data matrices (\code{X} and \code{Y}) 
 behind the scenes. Yet all of them provide a unified interface.
The methods below are constructors for the differnt backend 
objects and return an S3 object which implements pvalue 
methods listed above.
}
\details{
Given a testing set B, return the p-values for the opposite side.
This is an S3 method that is implemented by all backends.
}
\section{Methods (by class)}{
\itemize{
\item \code{perm}: Implementation for sum of squared correlations under independent Gene and SNP sets. 
Uses the permutation moments approximation.
}}

\keyword{internal}
